logstash/requirements.txt
python-evtx==0.6.1